requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.0.1
requests==2.31.0
google-genai>=1.74.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase>=2.10.0